% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot MDS for DESeq2 object}
\usage{
plotMDS(rld)
}
\arguments{
\item{rld}{A \code{\link{DESeqTransform}} object.}
}
\value{
A ggplot2 object
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
plotMDS(rld)
}
\description{
plotMDS function with ggplot2
}
\author{
Marisol Alvarez-Martinez
}
